Moved normalization and smoothing to the Process class.
