+++ skhyper/process/__init__.py
@@ -1,5 +1,3 @@
 from skhyper.process._process import Process
-from skhyper.process._normalization import normalize
-from skhyper.process._smoothing import savgol_filter
 
-__all__ = ['Process', 'normalize', 'savgol_filter']
+__all__ = ['Process']
