+++ skhyper/process/_process.py
@@ -4,6 +4,7 @@ Stores data in a custom class and generates attributes for other modules
 
 import warnings
 import numpy as np
+from scipy.signal import savgol_filter as _savgol
 
 from skhyper.view import hsiPlot
 from sklearn.decomposition import PCA
@@ -22,6 +23,9 @@ class Process:
 
     scale : bool
         Scales the spectra to either between {0, 1} or {-1, 1} depending on presence of negative values.
+
+    normalize : bool
+        Normalizes each spectra by subtracing the mean spectrum of the hyperspectral dataset.
 
     Attributes
     ----------
@@ -71,9 +75,10 @@ class Process:
     100000
 
     """
-    def __init__(self, X, scale=True):
+    def __init__(self, X, scale=True, normalize=False):
         self.data = X
         self._scale = scale
+        self._normalize = normalize
 
         # Data properties
         self.shape = None
@@ -102,6 +107,7 @@ class Process:
         # Perform data operations
         self._data_checks()
         if self._scale: self._data_scale()
+        if self._normalize: self._data_normalization()
         self._data_mean()
         self._data_access()
 
@@ -151,9 +157,29 @@ class Process:
 
         Scales the hyperspectral data to between 0 and 1 for all positive data or
         -1 and 1 for positive and negative data.
-
         """
         self.data = self.data / np.max(np.abs(self.data))
+
+    def _data_normalization(self):
+        """Normalize the hyperspectral data
+
+        Normalizes the hyperspectral data by subtracting the mean spectrum of the
+        data from each pixel.
+        """
+        if self.ndim == 3:
+            mean_spectrum = np.squeeze(np.mean(np.mean(self.data, 1), 0))
+
+            for xpix in range(self.shape[0]):
+                for ypix in range(self.shape[1]):
+                    self.data[xpix, ypix, :] -= mean_spectrum
+
+        elif self.ndim == 4:
+            mean_spectrum = np.squeeze(np.mean(np.mean(np.mean(self.data, 2), 1), 0))
+
+            for xpix in range(self.shape[0]):
+                for ypix in range(self.shape[1]):
+                    for zpix in range(self.shape[2]):
+                        self.data[xpix, ypix, zpix, :] -= mean_spectrum
 
     def _data_access(self):
         self.image = _AccessImage(self.data, self.shape, self.ndim)
@@ -165,6 +191,65 @@ class Process:
         Opens a hyperspectral viewer with the hyperspectral array loaded (pyqt GUI)
         """
         hsiPlot(self)
+
+    def smoothen(self, window_length=5, polyorder=3, deriv=0, delta=1.0, axis=-1, mode='interp', cval=0.0):
+        """ Savitzky-Golay filter
+
+        Applies a Savitzky-Golay filter to all the spectral components in the hyperspectral array.
+        The data is modified directly in the object, i.e. a new object is not created
+        (this is to limit memory usage).
+
+        Parameters
+        ----------
+        window_length : int, optional (default: 5)
+            The length of the filter window (i.e. the number of coefficients). `window_length`
+            must be a positive odd integer.
+
+        polyorder : int, optional (default: 3)
+            The order of the polynomial used to fit the samples. `polyorder` must be less than
+            `window_length`.
+
+        deriv : int, optional (default: 0)
+            The order of the derivative to compute. This must be a nonnegative intefer. The
+            default it 0, which means to filter the data without differentiating.
+
+        delta : float, optional (default: 1.0)
+            The spacing of the samples to which the data will be applied. This is only used
+            if `deriv`>0.
+
+        axis : int, optional (default: -1)
+            The axis of the array along which the filter is to be applied
+
+        mode : str, optional (default: 'interp')
+            Must be ‘mirror’, ‘constant’, ‘nearest’, ‘wrap’ or ‘interp’. This determines
+            the type of extension to use for the padded signal to which the filter is
+            applied. When mode is ‘constant’, the padding value is given by `cval`.
+            See the Notes for more details on ‘mirror’, ‘constant’, ‘wrap’, and
+            ‘nearest’. When the ‘interp’ mode is selected (the default), no
+            extension is used. Instead, a degree `polyorder` polynomial is fit to the
+            last `window_length` values of the edges, and this polynomial is used to
+            evaluate the last `window_length` // 2 output values.
+
+        cval : scalar, optional (default: 0.0)
+            Value to fill past the edges of the input if `mode` is 'constant'.
+
+        """
+        if self.ndim == 3:
+            for xpix in range(self.shape[0]):
+                for ypix in range(self.shape[1]):
+                    self.data[xpix, ypix, :] = _savgol(self.data[xpix, ypix, :], window_length=window_length,
+                                                       polyorder=polyorder, deriv=deriv, delta=delta,
+                                                       axis=axis, mode=mode, cval=cval)
+
+        elif self.ndim == 4:
+            for xpix in range(self.shape[0]):
+                for ypix in range(self.shape[1]):
+                    for zpix in range(self.shape[2]):
+                        self.data[xpix, ypix, zpix, :] = _savgol(self.data[xpix, ypix, :], window_length=window_length,
+                                                                 polyorder=polyorder, deriv=deriv, delta=delta,
+                                                                 axis=axis, mode=mode, cval=cval)
+
+        self.update()
 
     def flatten(self):
         """Flatten the hyperspectral data
